+++ sparselm/model/_base.py
