+++ sparselm/model/_ols.py
@@ -4,7 +4,7 @@ __author__ = "Fengyu Xie, Luis Barroso-Luque"
 
 import cvxpy as cp
 
-from .base import CVXEstimator
+from ._base import CVXEstimator
 
 
 class OrdinaryLeastSquares(CVXEstimator):
