+++ tests/test_optimizers.py
@@ -7,7 +7,7 @@ import cvxpy as cp
 import numpy as np
 import pytest
 
-from sparselm.model.miqp.regularized_l0 import L1L0, L2L0
+from sparselm.model.miqp._regularized_l0 import L1L0, L2L0
 from sparselm.optimizer import GridSearch, LineSearch
 
 ALL_CRITERION = ["max_r2", "one_std_r2"]
